// src/types/anime.ts
export interface AnimeCard {
  id: number;
  title: string;
  image: string;
  score: number | null;
  episodes: number | null;
  status: string;
  genres: string[];
  synopsis: string | null;
  year: number | null;
  trailer?: string;
}

export interface MangaCard {
  id: number;
  title: string;
  image: string;
  score: number | null;
  chapters: number | null;
  volumes: number | null;
  status: string;
  genres: string[];
  synopsis: string | null;
  authors: string[];
}